small fix for AC-1352

+++ kubedock/updates/kdmigrations/versions/56f9182bf415_add_persistent_disk.py
@@ -22,7 +22,7 @@ def upgrade():
         sa.Column('drive_name', sa.String(length=64), nullable=False),
         sa.Column('name', sa.String(length=64), nullable=False),
         sa.Column('owner_id', sa.Integer(), nullable=False),
-        sa.Column('size', sa.BigInteger(), nullable=False),
+        sa.Column('size', sa.Integer(), nullable=False),
         sa.Column('pod_id', postgresql.UUID(), nullable=True),
         sa.ForeignKeyConstraint(['owner_id'], ['users.id'], ),
         sa.ForeignKeyConstraint(['pod_id'], ['pods.id'], ),
